pyproject.toml
[tool.poetry]
name = "goit-web-pd-homework-6-database"
version = "0.1.0"
description = "Homework with database"
authors = ["pavlodubovyi <133531274+pavlodubovyi@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
faker = "^24.3.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
